+++ latch/registry/utils.py
@@ -19,6 +19,7 @@ from latch.registry.upstream_types.types import (
 from latch.registry.upstream_types.values import DBValue
 from latch.types.directory import LatchDir
 from latch.types.file import LatchFile
+from latch.utils import current_workspace
 
 # todo(maximsmol): hopefully, PyLance eventually narrows `TypedDict`` unions using `in`
 # then we can get rid of the casts
@@ -326,7 +327,7 @@ def to_registry_literal(
                 "cannot convert non-blob python literal to registry blob"
             )
 
-        ws_id = user_config.workspace_id
+        ws_id = current_workspace()
         if ws_id == "":
             ws_id = None