+++ latch_cli/services/get_executions.py
@@ -10,8 +10,8 @@ from apscheduler.schedulers.background import BackgroundScheduler
 from latch_sdk_config.latch import config
 
 import latch_cli.menus as menus
+from latch.utils import account_id_from_token, current_workspace, retrieve_or_login
 from latch_cli.tinyrequests import post
-from latch_cli.utils import account_id_from_token, current_workspace, retrieve_or_login
 
 
 def get_executions():