+++ CHANGELOG.md
@@ -16,6 +16,46 @@ Types of changes
 
 # Latch SDK Changelog
 
+## 2.40.2 - 2024-04-09
+
+### Fixed
+
+* latch register failing when using launchplans with file or directory types
+
+## 2.40.1 - 2024-04-06
+
+### Added
+
+* add support for allocating task resources at runtime
+
+## 2.40.0 - 2024-04-05
+
+### Fixed
+
+* latch commands default to the user's default workspace instead of personal workspace
+
+## 2.39.7 - 2024-04-03
+
+### Documentation
+
+* latch cp
+  + Commandline docstring states recursive default behavior.
+
+## 2.39.6 - 2024-04-02
+
+### Fixed
+
+* latch sync
+  + prints error when case other than local -> remote is attempted
+
+
+## 2.39.4 - 2024-03-11
+
+### Added
+
+* Snakemake
+  + Add `container_args` field in `EnvironmentConfig`
+
 ## 2.39.3 - 2024-03-07
 
 ### Fixed